pyproject.toml
[project]
name = "pick-and-go"
version = "0.1.0"
description = "Pick and Go backend python server's dependencies."
readme = "README.md"
requires-python = ">=3.10.17"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
